chore: use Badge from signoz for VolumeControlBadge
